Changed link with original label to account for double labels )
As a consequence of double labels in the tree (for example by cleaning up strains), the size of the table with matches was not equal to the original tip label vector and could not be automatically merged. An additional dataframe with original tip labels and lowercase search strings is generated and used as linking table.